youtube_uploader/requirements_youtube.txt
google-auth-oauthlib
google-api-python-client 
gspread
python-dotenv
pytz